fix: resolve merge conflicts
